DiskImageBuilder remove old folder and files
Changed the image_functions.sh to look for the right directory which
is "$IMAGE_NAME.d" and remove this. Added a check to remove old *.qcow2
files from previous builds. Confirmed the instance and image cleanup is
deleting these using the uuid in the db and not by their name. Added the
IMAGE_VISIBILITY var to all image_create as this needs setting when we
make new images private when testing.i Fixed TEST_IMAGE_NAME var as it
was causing the script to fail, as there was an image and instance using
the same names. Added a redirect of STDERR as this was causing the
delete files in the cleanup function to fail.