Let the nvidia-cuda element install cuda
Do not install cuda as part of packages.

Per

https://github.com/CSCfi/diskimage-builder/blob/csc/lib/img-functions#L143

the packages are installed prior to the elements getting
processed. If we take a head start installing cuda, this can fudge
up the whole kernel/headers installation part. So omit this package
installation altogether, and just let the nvidia-cuda element install
cuda at a more suitable time.