Remove image size specification
Currently, images are set to be created with a size explicitly set to 10 GB. From https://docs.openstack.org/diskimage-builder/latest/user_guide/building_an_image.html

--image-size
The size of loopback device which the image will be generated in, in gigabytes. If this is left unset, the size will be calculated from the on-disk size of the image and then scaled up by a fixed 60% factor.

This commit removes the size specification, so that the size of the image is automatically computed using the above rule.