Change image format from qcow2 to raw in cpouta devel (#44)
* Change image format from qcow2 to raw

The commit not only switches the format from qcow2 to raw, but makes it
so that it should be easy to switch between the two formats by just
changing the image format in the constants file.

* Remove image size specification

Currently, images are set to be created with a size explicitly set to 10 GB. From https://docs.openstack.org/diskimage-builder/latest/user_guide/building_an_image.html

--image-size

The size of loopback device which the image will be generated in, in gigabytes. If this is left unset, the size will be calculated from the on-disk size of the image and then scaled up by a fixed 60% factor.

This commit removes the size specification, so that the size of the image is automatically computed using the above rule.

With the current version of the tool it is not possible to specify --image-extra-space, which would allow to tweak the percentage value, so we need to be happy as-is for now.