Add logging of image building process output
cleanup is a single function that is called regardless if the image
building process is successful or if it fails. This commits adds a
distinction between the trap for the EXIT signal and the trap for the
ERR signal, which will print a different message. The original cleanup
function is called in both cases.