Add monitoring to image building process (#51)
This commit adds the basic tooling for monitoring the output of the image building process. The cleanup function in the image building scripts is leveraged to log a summary string about the outcome of the image building, i.e., if the process was successful or not. A python script, combined with configuration files, is then used to parse the logs and return an output understandable by our monitoring tools.